feat: Inject debug ids into source and sourcemap files before uploading
