+++ tests/integration/_cases/sourcemaps/sourcemaps-upload-successfully-upload-file.trycmd
@@ -2,6 +2,7 @@
 $ sentry-cli sourcemaps upload tests/integration/_fixtures/bundle.min.js.map --release=wat-release
 ? success
 > Found 1 release file
+
 > Analyzing 1 sources
 > Rewriting sources
 > Adding source map references
